Move some libraries back to shared
The Robolectric_* libraries are included into Robolectric_all.
Moving all the dependencies caused some additional classes to
show up in Robolectric_all, which caused RunRoboSettingsTests
to fail in AOSP, but didn't affect master.  Move the dependencies
for the libraries back to libs, and only leave the static
dependencies for the tests and the annotation processor.

Bug: 122331577
Test: m RunSettingsRoboTests Run_robolectric_test_suite
Test: diff packages in Robolectric_all before and after
Change-Id: I810264b68a881e73658bb1d3379c0fff1ab65879