Remove jars from APEXes and use framework-all instead.
The jars from APEXes are not available to outside of the APEXes.
Therefore they are removed from the dependencies. Instead, the
'framework' module is replaced with 'framework-all' which is the
build-only module built with the entire framework source files -
including the files for APEXes as well.

Bug: N/A
Test: m
Test: m -j RunFrameworksServicesRoboTests

Change-Id: I8c4d728aaf648eb15f933c8e13f38095ceca9bbc
Merged-In: I8c4d728aaf648eb15f933c8e13f38095ceca9bbc