Add shadow methods for clearing cross-profile app-ops
Add ShadowCrossProfileAppsTest#clearInteractAcrossProfilesAppOps and
unit tests.

Also add other missing shadow methods. Tests were not added for shadow
changes in ShadowUserManager since it's so out of sync with the
Robolectric source of truth currently that the tests would use different
fields. Unit tests in Robolectric in the Android source tree are not
currently working.

Fixes: 151145623
Test: atest com.android.managedprovisioning.task.CreateManagedProfileTaskRoboTest
Test: robolectric unit tests in Android source are broken but the new
unit tests were tested on the robolectric source of truth

Change-Id: I0e2b00586688a4dcd0d7e20a1cd1e39f6236794c